rename package to grants_tagger_light
