+++ grants_tagger_light/cli.py
@@ -7,11 +7,11 @@ import typer
 logger = logging.getLogger(__name__)
 
 
-from grants_tagger.predict import predict_cli
-from grants_tagger.evaluation import evaluate_app
-from grants_tagger.preprocessing import preprocess_app
-from grants_tagger.tune_threshold import tune_threshold_cli
-from grants_tagger.download_epmc import download_epmc_cli
+from grants_tagger_light.predict import predict_cli
+from grants_tagger_light.evaluation import evaluate_app
+from grants_tagger_light.preprocessing import preprocess_app
+from grants_tagger_light.tune_threshold import tune_threshold_cli
+from grants_tagger_light.download_epmc import download_epmc_cli
 
 
 app = typer.Typer()
