+++ grants_tagger_light/preprocessing/preprocess_mesh.py
@@ -14,9 +14,9 @@ from typing import Optional
 import pandas as pd
 import typer
 
-from grants_tagger.utils import write_jsonl, verify_if_paths_exist
-from grants_tagger.label_binarizer import create_label_binarizer
-from grants_tagger.split_data import split_data
+from grants_tagger_light.utils import write_jsonl, verify_if_paths_exist
+from grants_tagger_light.label_binarizer import create_label_binarizer
+from grants_tagger_light.split_data import split_data
 
 
 def yield_raw_data(input_path):
