+++ grants_tagger_light/preprocessing/preprocess_wellcome.py
@@ -14,9 +14,9 @@ import shutil
 
 import pandas as pd
 from typing import Optional
-from grants_tagger.utils import verify_if_paths_exist
-from grants_tagger.label_binarizer import create_label_binarizer
-from grants_tagger.split_data import split_data
+from grants_tagger_light.utils import verify_if_paths_exist
+from grants_tagger_light.label_binarizer import create_label_binarizer
+from grants_tagger_light.split_data import split_data
 
 
 def process_tagger_data(tagger_data):
